requirements.txt
APScheduler==3.10.4
fastapi==0.105.0
firebase-admin==6.3.0
paho-mqtt==1.6.1
uvicorn==0.25.0
uuid0==0.2.7